1-create_file.c up

+++ file_io/1-create_file.c
@@ -9,7 +9,7 @@
  */
 int create_file(const char *filename, char *text_content)
 {
-	ssize_t fd, fwrite;
+	int fd, fwrite;
 	int len;
 
 	if (filename == NULL)
